refactor: split out the different providers into different folders
